Case .STP file
A few notes, removed provision for USB-C power connection. Am now powering this setup via a USB cable cut and modified to plug into the Rpi power and ground pins via a power bank. that cable passes through the provided pass through located nera the GPIO pins. Passthrough in line with Pi ethernet connector is for a short patch cable to give the EUD a wired connection to the bridge. Final port is obviously for the XPAH antenna. Screws were 2mm for pi and 3 mm for case and td-xpah mounts if I remember correctly.  Will get a better BOM at some point